SQL Script and Nav Bar Updates
Added a SQL script database/setup.sql. This allows a new user to create the relevant database and tables.

Updated README.md to instruct new users.

Editing nav bar to limit those not logged in to either Log In or Register. Previous limitations still in place so if you go to a URL for instance create-event not logged in, you will still be redirected